requirements.txt
django==4.2.7
djangorestframework==3.14.0
gunicorn==21.2.0
dj-database-url==2.1.0
psycopg2-binary==2.9.5
gevent